Show celltoolbar on hover in addition to focus.
Make it easier to reach for an action of a cell with one less click.

One drawback is that with vertical toolbar with 1 line cells it's ugly
as toolbars overlap. Also the area where the cell is considered
"hovered" doe not include the area of the toolbar that overflow from the
bottom of the cell if the cell is less than 3 lines.

Also question, is it worth doing the focused/hover in react while it can
just be a css display option that could also have animations ?