Default spawn options to stdio: ignore
Spawn options are now folded into a set of default options, using `Object.assign`
(not a full merge).

Note: this is a breaking change, particularly for Hydrogen which was providing
"loose" stdout and stderr as notification messages in Atom (I love having this
when I'm working with Spark in Hydrogen).

Fixes #24