feat: add menu/help/about option
I ended up implementing a non-trivial modal system for us in here. It's
not 100% of the way fleshed out, but the basic machinery is in place.

Nothing complicated, it's just a little verbose. The idea is that you
dispatch an action to set the current modal, and a connected
ModalController component is listening to that particular hunk of state.