Allow contentRef in kernel* actions.
Initially, I didn't think this was necessary (and we may be able to roll
it back), but the updateDisplayEpic listens to LAUNCH_KERNEL_SUCCESSFUL
and then calls updateDisplay, which needs the contentRef. It's likely
the case that we should decouple this information, but this at least
allows me to get a step further without too much effort.