Use forked version of ijavascript to include zeromq@5 prebuilds
Currently the repo is in a weird state since `ijavascript` doesn't rely on the latest version of `zeromq` that ships with Node 10 prebuilt binaries: https://github.com/n-riesco/jp-kernel/pull/8

This means that we're currently installing two different versions of `zeromq`. For now let's use the unpublished version that includes the latest binaries to unblock people.