Correctly use MathJax Queue to render math
Calling `MathJax.Hub.Typeset` executes the command right away and doesn't make use of the MathJax Queue wich leads to MathJax errors or broken rendering when opening a new notebook containing math. It would be great to make a bugfix release for the `@nteract/mathjax` package so we can start using it in Hydrogen.

This largely reverts #3431 and properly fixes the root cause of the problem.