Improved TS workflow [v2] (#3559)
* 2nd attempt at improved TS workflow

This version of the workflow largely eliminates the need to manually
build typescript files, except for when you are building a package for
release.

It works by requiring each typescript package to include an `index.ts`
(or equivalent for any other public entry point) at the *root* of each
package. This file typically will just re-export src.

With this file in place, TypeScript can now automatically discover and
use dependent packages without needing any kind of compile step.

TS files are only compiled as part of a webpack build or if you run
`yarn build:packages`.

As an additional benefit, this workflow is simpler now. You no longer
need to manually maintain references between packages in the tsconfig
files.

Note that there is draft (incorrect) documentation for the TS workflow
here as well. It needs to be updated before this is merged.

Other Changes:

- designated the top-level `types` folder for type declaration. All
  flow types were consolidateed to flow-typed.
- brings back individual build tasks for each project, TS projects just
  run tsc. This makes everything more consistent
- eliminated all tsc-specific tasks from the top-level package
- added `build:clean` tasks everywhere and added a top-level task of
  the same name. This way you can eliminate all build artifacts without
  having to fully reinstall all dependencies.

- stricter types for messaging
- add type stubs required for different projects to build

* Add reminder to update docs

* Some files were building into the wrong location

* Build CI packages in series

CI appears to be failing due to memory issues. Building in series
will take longer but hopefully less prone to failure.

* Update TypeScript docs